Made the tests able to retrieve the updated one-time 2FA password without restarting Cypress! (this is yuge)
