+++ package.json
@@ -9,9 +9,10 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "child_process": "^1.0.2"
+    
   },
   "devDependencies": {
+    "child_process": "^1.0.2",
     "cypress": "^12.3.0",
     "typescript": "^4.9.3"
   }
